requirements.txt
py-trees==0.8.3
networkx==2.2
Shapely==1.7.1
psutil
xmlschema==1.0.18
ephem
tabulate
numpy
matplotlib
six
simple-watchdog-timer
carla==0.9.13
pygame
pyopenal
keyboard
./mediapipe-0.8.11-cp37-cp37m-win_amd64.whl
opencv-python==4.11.0.86
pandas 
openpyxl